requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0.0
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt,argon2]
python-dotenv
pytest
httpx
